fix(feat.hqnow): refactor HQNow helpers and publisher filter IDs
Move GraphQL, pagination, status parsing, and URL normalization
helpers into a new `helpers` module and update call sites.

Replace hardcoded publisher name-to-ID mapping with explicit `ids` in
`filters.json` and parse selected publisher IDs directly. Also hide
filters while searching via source config.
