+++ sources/pt.hqnow/res/filters.json
@@ -59,6 +59,57 @@
 			"Vertigo",
 			"Virgin Comics",
 			"WildStorm"
+		],
+		"ids": [
+			"",
+			"45",
+			"52",
+			"50",
+			"7",
+			"19",
+			"43",
+			"51",
+			"12",
+			"41",
+			"26",
+			"4",
+			"6",
+			"1",
+			"30",
+			"24",
+			"16",
+			"29",
+			"27",
+			"42",
+			"39",
+			"20",
+			"34",
+			"56",
+			"21",
+			"55",
+			"17",
+			"57",
+			"14",
+			"8",
+			"5",
+			"53",
+			"40",
+			"37",
+			"3",
+			"18",
+			"49",
+			"15",
+			"22",
+			"10",
+			"38",
+			"36",
+			"46",
+			"25",
+			"48",
+			"32",
+			"13",
+			"47",
+			"2"
 		]
 	}
 ]
